refactor: recipients => paper 로 네이밍 변경
