+++ src/entities/paper/@x/message.ts
