+++ src/entities/paper/index.ts
@@ -1,0 +1,3 @@
+export type { RecipientsResponse, Recipient, RecipientsParams } from "./paper.types"
+export { RECIPIENTS_LIMIT, RECIPIENTS_OFFSET, RECIPIENTS_TYPE } from "./paper.constants"
+export { recipientQueryKey } from "./paper-query-key"
