+++ src/entities/paper/paper-query-key.ts
@@ -1,0 +1,7 @@
+import type { PaperParams } from "./paper.types"
+
+export const paperQueryKey = {
+  all: ["papers"],
+  popularLists: (params: PaperParams) => [...paperQueryKey.all, "popular", params],
+  recentLists: (params: PaperParams) => [...paperQueryKey.all, "recent", params]
+} as const
