+++ src/entities/paper/paper.constants.ts
@@ -1,0 +1,3 @@
+export const PAPER_LIMIT = 8
+
+export const PAPER_OFFSET = 0
