--- src/entities/recipient/index.ts
@@ -1,3 +1,0 @@
-export type { RecipientsResponse, Recipient, RecipientsParams } from "./recipient.types"
-export { RECIPIENTS_LIMIT, RECIPIENTS_OFFSET, RECIPIENTS_TYPE } from "./recipient.constants"
-export { recipientQueryKey } from "./recipient-query-key"
