--- src/entities/recipient/recipient.constants.ts
@@ -1,8 +1,0 @@
-export const RECIPIENTS_LIMIT = 8
-
-export const RECIPIENTS_OFFSET = 0
-
-export const RECIPIENTS_TYPE = {
-  RECENT: "recent",
-  POPULAR: "popular"
-} as const
